Migrate from Oracle to Postgres for the production database
Both Oracle Cloud's Always Free tier and Cloudflare Tunnel/Zero Trust
(a self-hosting workaround considered in between) turned out to
require a credit card for identity verification even on their free
tiers. Drops the Oracle-only constraint entirely: rewrites all five
Flyway migrations to plain PostgreSQL dialect, swaps the JDBC driver
and Flyway module, switches local dev/test to H2's PostgreSQL
compatibility mode, and updates docker-compose.yml to a Postgres
container for local parity runs.

The CD runbook now documents Supabase's free tier (confirmed no card
required) in place of Oracle ADB + the tunnel workaround -- and is
simpler as a result, since a managed Postgres host is reachable
directly over the public internet with no tunnel needed at all.

Verified against both H2 (PostgreSQL mode, the default dev/CI path)
and a real local PostgreSQL 16 instance. The real-Postgres run caught
two things H2's emulation didn't: Flyway needs flyway-database-postgresql
explicitly (not bundled in flyway-core for this version), and
Scenario.paramsJson's @Lob annotation forced Postgres's large-object
API (which requires a transaction, unlike auto-commit H2/Oracle CLOB)
-- removed, since a plain TEXT column needs no special mapping.

+++ backend/src/main/resources/db/migration/V4__streaks.sql
@@ -1,7 +1,7 @@
 -- Daily usage streak: tracked per user (real calendar days), independent of
--- any single scenario account — this is a habit-forming mechanic, not a
+-- any single scenario account -- this is a habit-forming mechanic, not a
 -- trading-performance one (see the existing HOT_STREAK_5 achievement for that).
 
 ALTER TABLE users ADD last_active_date DATE;
-ALTER TABLE users ADD current_streak NUMBER(6) DEFAULT 0 NOT NULL;
-ALTER TABLE users ADD longest_streak NUMBER(6) DEFAULT 0 NOT NULL;
+ALTER TABLE users ADD current_streak INTEGER DEFAULT 0 NOT NULL;
+ALTER TABLE users ADD longest_streak INTEGER DEFAULT 0 NOT NULL;
